aegis-backend/requirements.txt
# =============================================================================
# Aegis AI Content Moderation Service - Dependencies
# =============================================================================

# Core Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Request/Response Handling
python-multipart>=0.0.9
pydantic>=2.5.0
ffmpeg-python>=0.2.0
httpx>=0.26.0

# =============================================================================
# AI/ML Libraries (Placeholders for Phase 2)
# Uncomment these when implementing actual AI models
# =============================================================================

torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.9.0
numpy>=1.26.0
pillow>=10.2.0
ultralytics>=8.2.0
openai-whisper>=20231117

# =============================================================================
# Development & Testing
# =============================================================================

pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0

# =============================================================================
# Optional: Production Deployment
# =============================================================================

# gunicorn>=21.2.0
# python-json-logger>=2.0.0
